move LinearPredictionGuider
