+++ examples/stable_diffusion_xl/gm/modules/diffusionmodules/sigma_sampling.py
@@ -12,7 +12,7 @@ class EDMSampling(nn.Cell):
         self.p_mean = p_mean
         self.p_std = p_std
 
-    @ms.jit  # ops.randn causes 'dynamic shape' error in PyNative mode
+    @ms.jit  # FIXME: ops.randn causes 'dynamic shape' error in PyNative mode
     def construct(self, n_samples):
         log_sigma = self.p_mean + self.p_std * ops.randn((n_samples,))
         return log_sigma.exp()
